refactor: rename Api class to Otf for better clarity fix: correct import paths for User and Otf classes docs: update example scripts to use Otf class instead of Api
refactor(bookings.py): rename Api to Otf to reflect updated class name
test(test_api.py): update import and usage of Api to Otf to match refactor
