+++ examples/workout_examples.py
@@ -1,14 +1,14 @@
 import asyncio
 import os
 
-from otf_api import Api
+from otf_api import Otf
 
 USERNAME = os.getenv("OTF_EMAIL")
 PASSWORD = os.getenv("OTF_PASSWORD")
 
 
 async def main():
-    otf = await Api.create(USERNAME, PASSWORD)
+    otf = await Otf.create(USERNAME, PASSWORD)
 
     resp = await otf.get_member_lifetime_stats()
     print(resp.model_dump_json(indent=4))
