+++ tests/test_api.py
@@ -1,8 +1,8 @@
 import pytest
-from otf_api.api import Api
+from otf_api.api import Otf
 
 
 @pytest.mark.asyncio
 async def test_api_raises_error_if_no_username_password():
     with pytest.raises(TypeError):
-        await Api.create()
+        await Otf.create()
